+++ Changes.md
@@ -3,6 +3,7 @@
 ## 0.9.1
 
 - `Decimal128` support
+- Relax `Sized` requirement for `SchemaLike::from_samples(..)`,  `SchemaLike::from_type(..)`, `SchemaLike::from_value(..)`
 
 ## 0.9.0